Fix order of mint note record commitments
- The added mint e2e test fails without the fix.
- Move key generation to test_utils.rs
